修改主键泛型为Int类型

+++ src/main/java/com/lord/advertis/dao/CategoryDao.java
@@ -3,6 +3,6 @@ package com.lord.advertis.dao;
 import com.lord.advertis.base.BaseMapper;
 import com.lord.advertis.domain.Category;
 
-public interface CategoryDao extends BaseMapper<Category, Long>{
+public interface CategoryDao extends BaseMapper<Category, Integer>{
 
 }
